Decoupled resource binder from pipeline
