+++ taichi/rhi/vulkan/CMakeLists.txt
@@ -9,6 +9,7 @@ target_sources(${VULKAN_RHI}
     vulkan_device.cpp
     vulkan_event.cpp
     vulkan_pipeline.cpp
+    vulkan_resource_binder.cpp
     vulkan_surface.cpp
     vulkan_stream.cpp
     vulkan_device_creator.cpp
